Added new code

+++ Deadline_diary/main.c
@@ -5,11 +5,12 @@
 
 #define LINE_LENGTH 100
 
-// TODO Fuggvenypointer ahol lehet
 // Ascending order
 // Menu
 // Refactor
 // Split into files
+// free lancolt lista
+// esemenyek.txt atnevezese
 
 typedef struct Event{
     char *name;
@@ -18,6 +19,8 @@ typedef struct Event{
     char *description;
     struct Event *next;
 } Event;
+
+typedef Event* (*searchFunctionPtr)(Event*, char*);
 
 int get_line_length(char *name){
     int length = strlen(name);
@@ -91,37 +94,40 @@ Event *add_event(Event *list, char *name, char *time, char *location, char *desc
 }
 
 Event *search_by_name(Event *list, char *name){
+    Event *new_list = NULL;
     for (Event *move = list; move != NULL; move = move->next){
         if (strcmp(move->name, name) == 0)
-            return move;
+            new_list = add_event(new_list, move->name, move->time, move->location, move->description);
     }
-    return NULL;
+    return new_list;
 }
 
 Event *search_by_date(Event *list, char *date){
+    Event *new_list = NULL;
     for (Event *move = list; move != NULL; move = move->next){
         if (strcmp(move->time, date) == 0)
-            return move;
+            new_list = add_event(new_list, move->name, move->time, move->location, move->description);
     }
-    return NULL;
+    return new_list;
 }
 
 Event *search_by_location(Event *list, char *location){
+    Event *new_list = NULL;
     for (Event *move = list; move != NULL; move = move->next){
         if (strcmp(move->location, location) == 0)
-            return move;
+            new_list = add_event(new_list, move->name, move->time, move->location, move->description);
     }
-    return NULL;
+    return new_list;
 }
 
-Event *delete_event_by_name(Event *list, char *name){
+Event *delete_event(Event *list, char *name, char *date, char *location){
     Event *move = list;
     Event *move_behind = NULL;
 
     if (list == NULL)
         return NULL;
     while (move != NULL){
-        if (strcmp(move->name, name) == 0) {
+        if (strcmp(move->name, name) == 0 && strcmp(move->time, date) == 0 && strcmp(move->location, location) == 0) {
             Event *delete_item = move;
             move = move->next;
             if (move_behind == NULL)
@@ -138,28 +144,21 @@ Event *delete_event_by_name(Event *list, char *name){
     return list;
 }
 
-void modify_event_by_name(Event *list, char *original, char *changed){
-    Event *item = search_by_name(list, original);
-    if (item == NULL)
-        printf("The item does not exist!\n");
-    else
-        item->name = changed;
-}
-
-void modify_event_by_date(Event *list, char *original, char *changed){
-    Event *item = search_by_name(list, original);
-    if (item == NULL)
-        printf("The item does not exist!\n");
-    else
-        item->time = changed;
-}
-
-void modify_event_by_location(Event *list, char *original, char *changed){
-    Event *item = search_by_name(list, original);
-    if (item == NULL)
-        printf("The item does not exist!\n");
-    else
-        item->location = changed;
+void modify_event(Event *list, char *name, char *date, char *location, char *mit, char *mire){
+    for (Event *move = list; move != NULL; move = move->next){
+        if (strcmp(move->name, name) == 0 && strcmp(move->time, date) == 0 && strcmp(move->location, location) == 0){
+            if (strcmp(mit, "name") == 0)
+                move->name = mire;
+            else if (strcmp(mit, "date") == 0)
+                move->time = mire;
+            else if (strcmp(mit, "location") == 0)
+                move->location = mire;
+            else if (strcmp(mit, "description") == 0)
+                move->description = mire;
+            else
+                printf("The given input was incorrect!\n\n");
+        }
+    }
 }
 
 void print_events(Event *list){
@@ -223,22 +222,62 @@ void write_to_file(char* filename, Event *list){
     fclose(fp);
 }
 
+void freeList(Event *list){
+    Event *move = list;
+    do {
+        Event *tmp = move->next;
+        free(move);
+        move = tmp;
+    } while (move != NULL);
+}
+
 int main(void) {
     Event *events = NULL;
     events = create_list("esemenyek.txt", events);
-    events = add_event(events, "Csuklo kontroll", "2024.04.10. 17:15", "Budapest, 1095, Mester utca 45-49.", "Ferencvarosi szakrendelo");
-    events = delete_event_by_name(events, "KEG Kocsmakviz");
+    //events = add_event(events, "Csuklo kontroll", "2024.04.10. 17:15", "Budapest, 1095, Mester utca 45-49.", "Ferencvarosi szakrendelo");
+
+    // Delete event
+    /*
+    events = delete_event(events, "KEG Kocsmakviz", "2024/04/07 19:00", "Budapest, 1114, Orlay u. 1.");
     print_events(events);
-    modify_event_by_name(events, "Csuklo kontroll", "Csuklo kontrol");
+    */
+
+    // Modify event attribute
+    /*
     print_events(events);
-    print_list_by_year(events, "2024");
-    print_list_by_month(events, "04");
-    print_list_by_day(events, "02");
-    Event *event = search_by_name(events, "KEG Kocsmakviz");
-    if (event == NULL)
+    modify_event(events, "Csuklo kontroll", "2024/04/10 17:15", "Budapest, 1095, Mester utca 45-49.", "name", "Csuklo kontrol");
+    print_events(events);
+    */
+
+    //print_list_by_year(events, "2024");
+    //print_list_by_month(events, "04");
+    //print_list_by_day(events, "02");
+
+    // Search function testing section
+    /*
+    searchFunctionPtr fptr = search_by_name;
+    Event *search_event = fptr(events, "KEG Kocsmakviz");
+    if (search_event == NULL)
         printf("The element does not exist!\n");
     else
-        printf("Event name: %s\nEvent time: %s\nEvent location: %s\nEvent description: %s\n", event->name, event->time, event->location, event->description);
-    write_to_file("esemenyek.txt", events);
+        print_events(search_event);
+    fptr = search_by_date;
+    search_event = fptr(events, "2024/04/12 16:20");
+    if (search_event == NULL)
+        printf("The element does not exist!\n");
+    else
+        print_events(search_event);
+    fptr = search_by_location;
+    search_event = fptr(events, "Budapest, 1095, Mester utca 45-49.");
+    if (search_event == NULL)
+        printf("The element does not exist!\n");
+    else
+        print_events(search_event);
+    */
+
+    //write_to_file("esemenyek.txt", events);
+
+    // Release the linked list
+    freeList(events);
     return 0;
 }
